+++ database/migrations/2022_08_20_104122_create_versions_table.php
@@ -13,6 +13,10 @@ return new class extends Migration
      */
     public function up()
     {
+        if (config('database.default') !== 'mysql_users') {
+            return;
+        }
+
         Schema::create('versions', function (Blueprint $table) {
             $table->id();
 
@@ -30,6 +34,10 @@ return new class extends Migration
      */
     public function down()
     {
+        if (config('database.default') !== 'mysql_users') {
+            return;
+        }
+
         Schema::dropIfExists('versions');
     }
 };
